consolidate redux and component code
