+++ client/react/frontpage/components/DJList.js
@@ -1,10 +1,18 @@
 // DJList.js
 
 import React from 'react';
+import { connect } from 'react-redux';
 import './DJList.scss';
 
 import DJInfo from './DJInfo';
 import Loader from './Loader';
+
+import {
+  updateDJs,
+  startFetching,
+  stopFetching,
+  fetchUpdatedDJs,
+} from '../actions/djs';
 
 const DJUrl = '/api/djs';
 
@@ -36,4 +44,15 @@ const DJList = React.createClass({
   },
 });
 
-export default DJList;
+const mapStateToProps = state => ({
+  djs: state.djs.djs,
+  fetching: state.djs.fetching,
+});
+
+const mapDispatchToProps = dispatch => ({
+  updateDJs: () => {
+    fetchUpdatedDJs(dispatch);
+  },
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(DJList);
