--- client/react/frontpage/containers/ShowsTab.js
@@ -1,24 +1,0 @@
-// ShowsTab.js
-// container for frontpage list of shows
-
-import { connect } from 'react-redux';
-
-import { fetchUpdatedShows } from '../actions/shows';
-import ShowsContent from '../components/ShowsContent';
-
-const mapStateToProps = state => ({
-  shows: state.shows.shows,
-  fetching: state.shows.fetching,
-  currentShowID: state.shows.nowPlayingID,
-  spotlightShowID: state.shows.spotlightID,
-});
-
-const mapDispatchToProps = dispatch => ({
-  updateShows: () => {
-    fetchUpdatedShows(dispatch);
-  },
-});
-
-const ShowsTab = connect(mapStateToProps, mapDispatchToProps)(ShowsContent);
-
-export default ShowsTab;
